[DEBT-043-A] Split cli/research.py into cli/research/ package
Refactored 1367-line god file into modular package (12 files, all ≤400 LoC):

- __init__.py (111): App registration, re-exports for backwards compat
- _shared.py (92): Shared utilities (_get_thesis_file, _load_theses, etc)
- context.py (186): Market context research command
- topic.py (123): Topic research command
- similar.py (84): Similar page search command
- deep.py (181): Deep research via Exa /research/v1
- backtest.py (202): Thesis backtesting command
- cache.py (37): Cache management commands
- thesis/__init__.py (60): Thesis subcommand router
- thesis/_commands.py (395): Core thesis CRUD commands
- thesis/_helpers.py (154): Render helpers for thesis display
- thesis/_suggest.py (58): Thesis suggestion command

Updated test patch paths to match new module structure.
All 42 research CLI tests pass.
