+++ docs/_debt/DEBT-043-srp-god-files.md
@@ -89,7 +89,7 @@ Adopt a strict size ceiling for `src/kalshi_research/**/*.py`:
 
 ## Acceptance Criteria (Phased)
 
-- [ ] Phase A: `cli/research.py` becomes `cli/research/` package and all files ≤400 lines
+- [x] Phase A: `cli/research.py` becomes `cli/research/` package and all files ≤400 lines
 - [ ] Phase B: `cli/scan.py` becomes `cli/scan/` package and all files ≤400 lines
 - [ ] Phase C: `api/client.py` split into endpoint modules; public import path preserved; all files ≤400 lines
 - [ ] Phase D: Remaining >400-line modules reduced under the ceiling
